% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.pcd_out}
\alias{autoplot.pcd_out}
\title{Plot convergence plots for pcd objects}
\usage{
\method{autoplot}{pcd_out}(x, tol = NULL, legend = TRUE)
}
\arguments{
\item{x}{A \code{pcd} object.}

\item{tol}{Tolerance for convergence. Default is \code{NULL},
which will use a default value of \code{1e-12}.}
}
\value{
A ggplot object.
}
\description{
Plot convergence plots for pcd objects
}
