% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcd.R
\name{cd_engine}
\alias{cd_engine}
\title{Proximal coordinate descent for penalized regression}
\usage{
cd_engine(
  X,
  y,
  lambda = 0,
  prox_fun = prox_soft,
  tol = 1e-08,
  max_iter = 1e+05,
  intercept_update = c("cyclic", "after_each_gradient", "after_each_newton",
    "after_each_exact", "after_sweep_gradient", "after_sweep_newton",
    "after_sweep_exact"),
  alpha = 1,
  verbose = FALSE
)
}
\arguments{
\item{X}{design matrix (n x p)}

\item{y}{response vector (n x 1)}

\item{lambda}{the penalty parameter (scalar). lambda >= 0.}

\item{prox_fun}{the proximal operator to use. Can be either "identity",
"soft", or "mcp". For "mcp", the \code{gamma} parameter must be specified.}

\item{tol}{the convergence tolerance. The algorithm stops when the maximum
change in coefficients is less than this value.}

\item{max_iter}{the maximum number of iterations. If -1, the algorithm
will run until convergence.}

\item{intercept_update}{the strategy for updating the intercept term.
\itemize{
\item cyclic: update the intercept term when the first predictor is visited.
\item after_each: update the intercept term after each slope update.
\item after_sweep_gradient: update the intercept term after each sweep a single
gradient descent step
\item after_sweep_newton: update the intercept term after each sweep a single
Newton's method step
\item after_sweep_exact: update the intercept term after using Newtons method
until convergence
}}

\item{verbose}{if TRUE, print progress messages.}
}
\value{
numeric vector of coefficients (p x 1), beta
}
\description{
Proximal coordinate descent for penalized regression
}
