% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcd.R
\name{pcd}
\alias{pcd}
\title{Proximal coordinate descent for penalized regression}
\usage{
pcd(
  formula,
  data,
  lambda,
  prox_fun = c("identity", "soft", "mcp"),
  intercept_update = c("cyclic", "after_each_gradient", "after_each_newton",
    "after_each_exact", "after_sweep_gradient", "after_sweep_newton",
    "after_sweep_exact"),
  family = c("gaussian", "binomial"),
  method = c("gradient", "newton"),
  tol = 1e-08,
  max_iter = 1e+05,
  standardize = TRUE,
  gamma = -1,
  verbose = FALSE
)
}
\arguments{
\item{formula}{an object of class "formula"
(or one that can be coerced to that class):
a symbolic description of the model to be fitted.}

\item{data}{an optional data frame, list or environment
containing the variables in the model.}

\item{lambda}{the penalty parameter (scalar). lambda >= 0.}

\item{prox_fun}{the proximal operator to use. Can be either "identity",
"soft", or "mcp". For "mcp", the \code{gamma} parameter must be specified.}

\item{intercept_update}{the strategy for updating the intercept term.
\itemize{
\item cyclic: update the intercept term when the first predictor is visited.
\item after_each_\*: update the intercept term after each slope update.
\item after_sweep_gradient: update the intercept term after each sweep a single
gradient descent step
\item after_sweep_newton: update the intercept term after each sweep a single
Newton's method step
\item after_sweep_exact: update the intercept term after using Newtons method
until convergence
}}

\item{family}{If "gaussian", the model is fitted using ordinary least squares.
If "binomial", the model is fitted using logistic regression.}

\item{method}{If family is "binomial", this argument specifies if
the method used for coordinate descent should be a first-order method
("gradient), using an upper bound of the Hessian, or a proximal Newton method.}

\item{tol}{the convergence tolerance. The algorithm stops when the maximum
change in coefficients is less than this value.}

\item{max_iter}{the maximum number of iterations. If -1, the algorithm
will run until convergence.}

\item{standardize}{if TRUE, standardize the predictors before fitting the model.}

\item{gamma}{the concavity parameter for MCP. This parameter is only
relevant if \code{prox_fun} is "mcp". It should be a positive number or \code{Inf},
which indicates that the MCP is equivalent to soft-thresholding.}

\item{verbose}{if TRUE, print progress messages.}
}
\value{
data.frame(term, estimate)
}
\description{
This function implements proximal coordinate descent for penalized regression
using the specified proximal operator. It supports both linear regression,
lasso regression, and logistic regression with the option to use
either soft-thresholding or minimax concave penalty (MCP) as the proximal
operator.
}
