% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcd.R
\name{prox_mcp}
\alias{prox_mcp}
\title{Minimax concave penalty proximal operator (MCP)}
\usage{
prox_mcp(gamma)
}
\arguments{
\item{gamma}{concavity parameter (default = Inf)
If gamma = Inf then MCP = soft-thresholding}

\item{z}{vector of coefficients}

\item{lambda}{penalty parameter}
}
\description{
Minimax concave penalty proximal operator (MCP)
}
\keyword{internal}
