% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcd.R
\name{prox_soft}
\alias{prox_soft}
\title{Soft-threshold proximal operator (ℓ₁ penalty)}
\usage{
prox_soft(z, lambda)
}
\arguments{
\item{z}{vector of coefficients}

\item{lambda}{penalty parameter. If lambda = 0, then this is the identity
operator.}
}
\description{
Soft-threshold proximal operator (ℓ₁ penalty)
}
\keyword{internal}
