% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_correlated}
\alias{simulate_correlated}
\title{Simulate correlated design}
\usage{
simulate_correlated(
  n,
  p,
  k,
  rho = 0,
  error_dist = c("normal", "t", "uniform"),
  beta = NULL,
  seed = 96
)
}
\arguments{
\item{n}{Number of observations}

\item{p}{Number of features}

\item{k}{Number of non-zero coefficients}

\item{rho}{Correlation coefficient used to simulate from uniform 0 to rho}

\item{error_dist}{Error distribution: "normal", "t", "uniform"}

\item{beta}{Optional true coefficient vector}

\item{seed}{Random seed}
}
\description{
Simulate correlated design
}
