% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_ar1}
\alias{simulate_ar1}
\title{Simulate VAR(1) / multivariate AR(1) design}
\usage{
simulate_ar1(
  n,
  p,
  k,
  rho = 0,
  burnin = 50,
  noise = c("mnormal", "mt"),
  sigma = NULL,
  df = 5,
  Psi = NULL,
  error_dist = c("normal", "t", "uniform"),
  beta = NULL,
  seed = 96
)
}
\arguments{
\item{n}{Number of observations}

\item{p}{Number of features}

\item{k}{Number of non-zero coefficients}

\item{rho}{Correlation coefficient for AR(1) process}

\item{burnin}{Number of initial observations to discard}

\item{noise}{Type of noise to use for AR process: "mnormal" or "mt"}

\item{sigma}{Covariance matrix for the noise}

\item{df}{Degrees of freedom for t-distribution for noise}

\item{Psi}{AR coefficient matrix}

\item{error_dist}{Error distribution: "normal", "t", "uniform"}

\item{beta}{Optional true coefficient vector}

\item{seed}{Random seed}
}
\description{
Simulate VAR(1) / multivariate AR(1) design
}
